pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["saatarko <saatar@tut.by>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
amqp = "5.2.0"
asgiref = "3.8.1"
billiard = "4.2.0"
build = "1.2.1"
cachecontrol = "0.14.0"
celery = "5.4.0"
certifi = "2024.6.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
crashtest = "0.4.1"
cryptography = "42.0.8"
defusedxml = "0.8.0rc2"
distlib = "0.3.8"
django = "4.2.13"
django-templated-mail = "1.1.1"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
djoser = "2.2.3"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
flower = "2.0.1"
humanize = "4.9.0"
idna = "3.7"
importlib-metadata = "8.2.0"
importlib-resources = "6.4.0"
installer = "0.7.0"
jaraco-classes = "3.4.0"
keyring = "24.3.1"
kombu = "5.3.7"
more-itertools = "10.4.0"
msgpack = "1.0.8"
oauthlib = "3.2.2"
packaging = "24.1"
pexpect = "4.9.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
prometheus-client = "0.20.0"
prompt-toolkit = "3.0.47"
ptyprocess = "0.7.0"
pycparser = "2.22"
pyjwt = "2.8.0"
pyproject-hooks = "1.1.0"
python-dateutil = "2.9.0.post0"
python3-openid = "3.2.0"
pytz = "2024.1"
pywin32-ctypes = "0.2.2"
rapidfuzz = "3.9.6"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
six = "1.16.0"
social-auth-app-django = "5.4.1"
social-auth-core = "4.5.4"
soupsieve = "2.5"
sqlparse = "0.5.0"
tomli = "2.0.1"
tomlkit = "0.13.0"
tornado = "6.4.1"
trove-classifiers = "2024.7.2"
typing-extensions = "4.12.2"
tzdata = "2024.1"
urllib3 = "2.2.2"
vine = "5.1.0"
virtualenv = "20.26.3"
wcwidth = "0.2.13"
zipp = "3.19.2"
redis = "^5.0.8"
drf-spectacular = {extras = ["swagger-ui"], version = "^0.27.2"}
django-celery-beat = "^2.6.0"
pillow = "^10.4.0"
django-cors-headers = "^4.4.0"
channels = "^4.1.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
websockets = "^12.0"
wsproto = "^1.2.0"
django-filter = "^24.3"
gunicorn = "^23.0.0"
black = "^24.8.0"
isort = "^5.13.2"
spectacular = "^0.4.0"
swagger-ui = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
